refactor: rename `RemotePollerConfig` to `RemoteConnectionConfig` for reuse with destinations
