+++ src/functions/ftp/external-poller/pollers/ftpPoller.ts
@@ -8,7 +8,7 @@ import { PutObjectCommand } from "@stedi/sdk-client-buckets";
 import { bucketClient } from "../../../../lib/buckets.js";
 import { FileDetails, ProcessingError, RemoteFileDetails } from "../types.js";
 import { DestinationBucket } from "../../../../lib/types/PartnerRouting.js";
-import { ConnectionDetails } from "../../../../lib/types/RemotePollerConfig.js";
+import { ConnectionDetails } from "../../../../lib/types/RemoteConnectionConfig.js";
 import { RemotePoller } from "./remotePoller.js";
 
 export class FtpPoller extends RemotePoller {
