+++ src/functions/ftp/external-poller/pollers/remotePoller.ts
@@ -1,6 +1,6 @@
 import { FileDetails, RemoteFileDetails } from "../types.js";
 import { DestinationBucket } from "../../../../lib/types/PartnerRouting.js";
-import { ConnectionDetails } from "../../../../lib/types/RemotePollerConfig.js";
+import { ConnectionDetails } from "../../../../lib/types/RemoteConnectionConfig.js";
 
 export abstract class RemotePoller {
   abstract getRemoteFileDetails(remotePath: string, remoteFiles?: string[]): Promise<RemoteFileDetails>;
